pyproject.toml
[tool.poetry]
name = "worldindex"
version = "0.1.0"
description = "WorldIndex"
authors = ["Rivaan Ranawat"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
faiss-cpu = { version = "^1.13.2", python = ">=3.11,<3.15" }
numpy = "^2.2.0"
pillow = "^11.1.0"
polars = "^1.26.0"
pydantic = "^2.10.0"
pyarrow = "^19.0.1"
pyyaml = "^6.0.2"
scikit-learn = "^1.8.0"
structlog = "^25.1.0"
fastapi = "^0.135.1"
uvicorn = "^0.41.0"
python-multipart = "^0.0.22"

[tool.poetry.group.video]
optional = true

[tool.poetry.group.video.dependencies]
lerobot = "*"
torch = "*"
torchcodec = "*"
transformers = "^4.57.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
httpx = "^0.28.1"

[tool.pytest.ini_options]
markers = [
  "integration: requires live LeRobot streaming access and the optional video dependency group",
]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
